QueryActionDisctoverPools now also returns (if available) the number of available units for each resource type. This value is returned as an integer attribute in the resource pool descriptor (See ResourceProperties). Updated the XmlRpcHandler to keep track not only of domain->Resource type but also of the available units for that resource type.
Extended the mapping to support one domain to multiple resource types. However, the rest of the code still assumes 
one-to-one mapping and is calling getDefatulResourceType() to achieve that.

Made discoverTypes take a parameter to say whether the call should wait for the response or not. This avoids the busywait loops that
I saw in the code (I removed these).